Refactor JuMP related code
